refactor: separate test
